requirements.txt
types-Pillow>=9.0.7
click>=7.1.2
Pillow>=8.2.0
